DS-131: Refactoring
