+++ ngx-fudis/projects/ngx-fudis/src/lib/components/form/error-summary/error-summary.service.ts
@@ -17,8 +17,6 @@ export class ErrorSummaryService {
 	updateErrorList(message: TFudisFormErrorSummaryItem) {
 		const currentErrors = this.errorListBus.value;
 
-		console.log(message);
-
 		const errorListIndex = currentErrors.findIndex((error) => error.id === message.id);
 
 		if (errorListIndex !== -1 && message.errors.length > 0) {
