+++ ngx-fudis/projects/ngx-fudis/src/lib/components/form/fieldset/fieldset.component.spec.ts
@@ -2,8 +2,9 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { MockComponent } from 'ng-mocks';
 import { FieldsetComponent } from './fieldset.component';
-import { GuidanceComponent } from '../guidance/guidance.component';
 import { ErrorSummaryComponent } from '../error-summary/error-summary.component';
+import { ErrorSummaryService } from '../error-summary/error-summary.service';
+import { GridComponent } from '../../grid/grid.component';
 
 describe('FieldsetComponent', () => {
 	let component: FieldsetComponent;
@@ -11,7 +12,8 @@ describe('FieldsetComponent', () => {
 
 	beforeEach(async () => {
 		await TestBed.configureTestingModule({
-			declarations: [FieldsetComponent, MockComponent(GuidanceComponent), MockComponent(ErrorSummaryComponent)],
+			declarations: [FieldsetComponent, MockComponent(GridComponent), MockComponent(ErrorSummaryComponent)],
+			providers: [ErrorSummaryService],
 		}).compileComponents();
 
 		fixture = TestBed.createComponent(FieldsetComponent);
