+++ ngx-fudis/projects/ngx-fudis/src/lib/components/form/input-with-language-options/input-with-language-options.component.scss
@@ -3,14 +3,13 @@
 .fudis-input-with-language-options {
 	box-sizing: border-box;
 	display: flex;
-	max-width: 100%;
 
 	&__l {
-		width: calc(23 * spacing.$spacing-s);
+		max-width: calc(23 * spacing.$spacing-s);
 	}
 
 	&__m {
-		width: calc(14 * spacing.$spacing-s);
+		max-width: calc(14 * spacing.$spacing-s);
 	}
 
 	&__s {
