+++ ngx-fudis/projects/ngx-fudis/src/lib/components/form/text-input/text-input.component.scss
@@ -8,18 +8,18 @@
 .fudis-text-input {
 	box-sizing: border-box;
 	display: block;
-	max-width: 100%;
+	width: 100%;
 
 	&__l {
-		width: calc(23 * spacing.$spacing-s);
+		max-width: calc(23 * spacing.$spacing-s);
 	}
 
 	&__m {
-		width: calc(14 * spacing.$spacing-s);
+		max-width: calc(14 * spacing.$spacing-s);
 	}
 
 	&__s {
-		width: calc(10 * spacing.$spacing-s);
+		max-width: calc(10 * spacing.$spacing-s);
 	}
 
 	&__input {
