+++ ngx-fudis/projects/ngx-fudis/src/lib/types/forms.ts
@@ -26,7 +26,7 @@ export type TFudisInputErrorMessages = {
 	matDatepickerParse?: string;
 };
 
-export type TFudisFieldsetErrorMessages = {
+export type TFudisGroupErrorMessages = {
 	[key: string]: TFudisInputErrorMessages;
 };
