+++ ngx-fudis/projects/ngx-fudis/src/public-api.ts
@@ -32,10 +32,7 @@ export {
 	ExpandableActionsDirective,
 } from './lib/components/expandable/expandable-directives';
 export { FieldsetComponent } from './lib/components/form/fieldset/fieldset.component';
-export {
-	FieldsetGuidanceDirective,
-	FieldsetContentDirective,
-} from './lib/components/form/fieldset/fieldset-directives';
+export { FieldsetContentDirective } from './lib/components/form/fieldset/fieldset-directives';
 export { InputWithLanguageOptionsComponent } from './lib/components/form/input-with-language-options/input-with-language-options.component';
 export { GridComponent } from './lib/components/grid/grid.component';
 export { GridDirective } from './lib/directives/grid/grid.directive';
